+++ source/main/index.js
@@ -11,6 +11,7 @@ import {
   generateWalletMigrationReport,
 } from './utils/setupLogging';
 import { handleDiskSpace } from './utils/handleDiskSpace';
+import { handleCheckBlockReplayProgress } from './utils/handleCheckBlockReplayProgress';
 import { createMainWindow } from './windows/main';
 import { installChromeExtensions } from './utils/installChromeExtensions';
 import { environment } from './environment';
@@ -164,6 +165,8 @@ const onAppReady = async () => {
   };
   mainErrorHandler(onMainError);
   await handleCheckDiskSpace();
+
+  await handleCheckBlockReplayProgress(mainWindow, launcherConfig.logsPrefix);
 
   cardanoNode = setupCardanoNode(launcherConfig, mainWindow);