+++ source/renderer/app/api/utils/patchAdaApi.js
@@ -18,7 +18,7 @@ import type {
   NetworkInfoResponse,
 } from '../network/types';
 import type { GetNewsResponse } from '../news/types';
-import { EPOCH_LENGTH_ITN } from '../../config/epochsConfig';
+import { getEpochLength } from '../../config/epochsConfig';
 
 let LATEST_APP_VERSION = null;
 let LOCAL_TIME_DIFFERENCE = 0;
@@ -46,6 +46,9 @@ export default (api: AdaApi) => {
           : get(sync_progress, 'quantity', 0);
 
       // extract relevant data before sending to NetworkStatusStore
+      const nextEpochNumber = get(next_epoch, 'epoch_number', null);
+      const nextEpochStartTime = get(next_epoch, 'epoch_start_time', '');
+      const epochLength = getEpochLength();
       return {
         syncProgress: SYNC_PROGRESS !== null ? SYNC_PROGRESS : syncProgress,
         localTip: {
@@ -53,19 +56,22 @@ export default (api: AdaApi) => {
           slot: get(node_tip, 'slot_number', 0),
         },
         networkTip: {
-          epoch: get(network_tip, 'epoch_number', 0),
-          slot: get(network_tip, 'slot_number', 0),
+          epoch: get(network_tip, 'epoch_number', null),
+          slot: get(network_tip, 'slot_number', null),
         },
         nextEpoch: {
-          epochNumber: get(next_epoch, 'epoch_number', 0),
-          epochStart: get(next_epoch, 'epoch_start_time', ''),
+          // N+1 epoch
+          epochNumber: nextEpochNumber,
+          epochStart: nextEpochStartTime,
         },
         futureEpoch: {
-          epochNumber: get(next_epoch, 'epoch_number', 0) + 1,
-          epochStart: moment(get(next_epoch, 'epoch_start', '')).add(
-            EPOCH_LENGTH_ITN,
-            'seconds'
-          ),
+          // N+2 epoch
+          epochNumber: nextEpochNumber ? nextEpochNumber + 1 : null,
+          epochStart: nextEpochStartTime
+            ? moment(nextEpochStartTime)
+                .add(epochLength, 'seconds')
+                .toISOString()
+            : '',
         },
       };
     } catch (error) {