+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnectingBackground.js
@@ -9,7 +9,6 @@ import styles from './SyncingConnectingBackground.scss';
 type Props = {
   hasLoadedCurrentTheme: boolean,
   isIncentivizedTestnet: boolean,
-  isShelleyTestnet: boolean,
   isConnecting: boolean,
   isSyncing: boolean,
 };
@@ -21,7 +20,6 @@ export default class SyncingConnectingBackground extends Component<Props> {
       isConnecting,
       isSyncing,
       isIncentivizedTestnet,
-      isShelleyTestnet,
       hasLoadedCurrentTheme,
     } = this.props;
     const componentStyles = classNames([
@@ -32,7 +30,7 @@ export default class SyncingConnectingBackground extends Component<Props> {
     ]);
     return (
       <div className={componentStyles}>
-        {isIncentivizedTestnet && !isShelleyTestnet && (
+        {isIncentivizedTestnet && (
           <>
             <div className={styles.backgroundOverlay} />
             <SVGInline