+++ source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsChooseWalletDialog.js
@@ -57,7 +57,23 @@ const messages = defineMessages({
     defaultMessage:
       '!!!This wallet does not contain the minimum amount of {minDelegationFunds} ADA which is required for delegation to be available. Please select a wallet with <span>a minimum amount of {minDelegationFunds} ADA</span> and click continue.',
     description:
-      'MinDelegationFunds Error Label on the delegation setup "choose wallet" step dialog.',
+      'errorMinDelegationFunds Error Label on the delegation setup "choose wallet" step dialog.',
+  },
+  errorMinDelegationFundsHasRewards: {
+    id:
+      'staking.delegationSetup.chooseWallet.step.dialog.errorMinDelegationFundsHasRewards',
+    defaultMessage:
+      '!!!This wallet does not contain the minimum amount of {minDelegationFunds} ADA which is required for delegation to be available.',
+    description:
+      'errorMinDelegationFundsHasRewards Error Label on the delegation setup "choose wallet" step dialog.',
+  },
+  errorMinDelegationFundsRewardsOnly: {
+    id:
+      'staking.delegationSetup.chooseWallet.step.dialog.errorMinDelegationFundsRewardsOnly',
+    defaultMessage:
+      '!!!This wallet contains only rewards balances so it cannot be delegated.',
+    description:
+      'errorMinDelegationFundsRewardsOnly Error Label on the delegation setup "choose wallet" step dialog.',
   },
   errorRestoringWallet: {
     id: 'staking.delegationSetup.chooseWallet.step.dialog.errorRestoringWallet',
@@ -126,12 +142,36 @@ export default class DelegationStepsChooseWalletDialog extends Component<
       getStakePoolById,
     } = this.props;
 
-    const selectedWallet: ?Wallet =
-      wallets.find(
-        (wallet: Wallet) => wallet && wallet.id === selectedWalletId
-      ) || null;
-    const amount = selectedWallet ? selectedWallet.amount : null;
-    const isAcceptableSetupWallet = amount && isWalletAcceptable(amount);
+    const selectedWallet: ?Wallet = wallets.find(
+      (wallet: Wallet) => wallet && wallet.id === selectedWalletId
+    );
+
+    const { amount, reward, isRestoring } = selectedWallet || {};
+
+    let errorMessage;
+    if (selectedWallet && !isWalletAcceptable(amount, reward)) {
+      // Wallet is restoring
+      if (isRestoring) errorMessage = messages.errorRestoringWallet;
+      // Wallet only has Reward balance
+      else if (!amount.isZero() && amount.equals(reward))
+        errorMessage = messages.errorMinDelegationFundsRewardsOnly;
+      // Wallet balance - rewards < min delegation funds
+      /*
+      else if (!reward.isZero())
+        errorMessage = messages.errorMinDelegationFundsHasRewards;
+      */
+      // Wallet balance < min delegation funds
+      else errorMessage = messages.errorMinDelegationFunds;
+    }
+
+    const error = errorMessage && (
+      <p className={styles.errorMessage}>
+        <FormattedHTMLMessage
+          {...errorMessage}
+          values={{ minDelegationFunds }}
+        />
+      </p>
+    );
 
     const dialogClassName = classNames([
       commonStyles.delegationSteps,
@@ -139,26 +179,9 @@ export default class DelegationStepsChooseWalletDialog extends Component<
     ]);
     const contentClassName = classNames([commonStyles.content, styles.content]);
 
-    const errorMinDelegationFunds = selectedWalletId &&
-      !isAcceptableSetupWallet && (
-        <p className={styles.errorMessage}>
-          <FormattedHTMLMessage
-            {...messages.errorMinDelegationFunds}
-            values={{ minDelegationFunds }}
-          />
-        </p>
-      );
-
-    const errorRestoringWallet = selectedWallet &&
-      selectedWallet.isRestoring && (
-        <p className={styles.errorMessage}>
-          {intl.formatMessage(messages.errorRestoringWallet)}
-        </p>
-      );
-
     const walletSelectClasses = classNames([
       styles.walletSelect,
-      errorMinDelegationFunds || errorRestoringWallet ? styles.error : null,
+      error ? styles.error : null,
     ]);
 
     const actions = [
@@ -167,10 +190,7 @@ export default class DelegationStepsChooseWalletDialog extends Component<
         label: intl.formatMessage(messages.continueButtonLabel),
         onClick: this.onSelectWallet.bind(this),
         primary: true,
-        disabled:
-          !selectedWalletId ||
-          !!errorMinDelegationFunds ||
-          !!errorRestoringWallet,
+        disabled: !selectedWalletId || !!error,
       },
     ];
 
@@ -223,8 +243,7 @@ export default class DelegationStepsChooseWalletDialog extends Component<
             value={selectedWalletId}
             getStakePoolById={getStakePoolById}
           />
-          {errorMinDelegationFunds}
-          {errorRestoringWallet}
+          {error}
         </div>
       </Dialog>
     );