+++ storybook/stories/nodes/syncing/SyncingConnecting.stories.js
@@ -3,10 +3,7 @@ import React from 'react';
 import { boolean, radios } from '@storybook/addon-knobs';
 import { linkTo } from '@storybook/addon-links';
 import { action } from '@storybook/addon-actions';
-import {
-  isIncentivizedTestnetTheme,
-  isShelleyTestnetTheme,
-} from '../../_support/utils';
+import { isIncentivizedTestnetTheme } from '../../_support/utils';
 
 import SyncingConnecting from '../../../../source/renderer/app/components/loading/syncing-connecting/SyncingConnecting';
 import { CardanoNodeStates } from '../../../../source/common/types/cardano-node.types';
@@ -17,6 +14,8 @@ export const DefaultSyncingConnectingStory = (props: {
   <SyncingConnecting
     hasNotification={false}
     hasUpdate={false}
+    isVerifyingBlockchain={false}
+    verificationProgress={0}
     hasUnreadAlerts={false}
     hasUnreadAnnouncements={false}
     hasUnreadNews={false}
@@ -28,7 +27,6 @@ export const DefaultSyncingConnectingStory = (props: {
     )}
     hasBeenConnected={boolean('hasBeenConnected', false)}
     isFlight={false}
-    isShelleyTestnet={isShelleyTestnetTheme(props.currentTheme)}
     isConnected={boolean('isConnected', false)}
     isSynced={boolean('isSynced', false)}
     isConnecting={boolean('isConnecting', true)}
@@ -63,6 +61,8 @@ export const ConnectivityIssuesSyncingConnectingStory = (props: {
   <SyncingConnecting
     hasNotification={false}
     hasUpdate={false}
+    isVerifyingBlockchain={false}
+    verificationProgress={0}
     hasUnreadAlerts={false}
     hasUnreadAnnouncements={false}
     hasUnreadNews={false}
@@ -72,7 +72,6 @@ export const ConnectivityIssuesSyncingConnectingStory = (props: {
     cardanoNodeState={CardanoNodeStates.RUNNING}
     hasBeenConnected
     isFlight={false}
-    isShelleyTestnet={isShelleyTestnetTheme(props.currentTheme)}
     isSynced={false}
     isConnecting
     isSyncing={false}
@@ -106,6 +105,8 @@ export const LoadingWalletDataSyncingConnectingStory = (props: {
   <SyncingConnecting
     hasNotification={false}
     hasUpdate={false}
+    isVerifyingBlockchain={false}
+    verificationProgress={0}
     hasUnreadAlerts={false}
     hasUnreadAnnouncements={false}
     hasUnreadNews={false}
@@ -114,7 +115,6 @@ export const LoadingWalletDataSyncingConnectingStory = (props: {
     cardanoNodeState={CardanoNodeStates.RUNNING}
     hasBeenConnected
     isFlight={false}
-    isShelleyTestnet={isShelleyTestnetTheme(props.currentTheme)}
     isSynced={false}
     isConnecting={false}
     isSyncing